refactored OperateImage class to take list
