+++ autogalaxy/imaging/fit_imaging.py
@@ -3,6 +3,7 @@ import numpy as np
 from autoconf import conf
 import autoarray as aa
 
+from autogalaxy.operate.image import OperateImage
 from autogalaxy.galaxy.galaxy import Galaxy
 from autogalaxy.plane.plane import Plane
 
@@ -46,7 +47,7 @@ class FitImaging(aa.FitImaging):
             image = dataset.image
             noise_map = dataset.noise_map
 
-        self.blurred_image = plane.blurred_image_2d_via_convolver_from(
+        self.blurred_image = self.calc_image.blurred_image_2d_via_convolver_from(
             grid=dataset.grid,
             convolver=dataset.convolver,
             blurring_grid=dataset.blurring_grid,
@@ -117,7 +118,9 @@ class FitImaging(aa.FitImaging):
     @property
     def model_images_of_galaxies(self):
 
-        model_images_of_galaxies = self.plane.blurred_images_of_galaxies_via_psf_from(
+        calc_image_list = OperateImageList(light_obj_list=self.plane.galaxies)
+
+        model_images_of_galaxies = calc_image_list.blurred_image_2d_list_via_psf_from(
             grid=self.grid,
             psf=self.imaging.psf,
             blurring_grid=self.imaging.blurring_grid,
@@ -156,13 +159,13 @@ class FitImaging(aa.FitImaging):
 
     @property
     def unmasked_blurred_image(self):
-        return self.plane.unmasked_blurred_image_2d_via_psf_from(
+        return self.calc_image.unmasked_blurred_image_2d_via_psf_from(
             grid=self.grid, psf=self.imaging.psf
         )
 
     @property
     def unmasked_blurred_image_of_galaxies(self):
-        return self.plane.unmasked_blurred_image_of_galaxies_via_psf_from(
+        return self.calc_image_list.unmasked_blurred_image_2d_list_via_psf_from(
             grid=self.grid, psf=self.imaging.psf
         )
